check_arg_mapper

Checking the arguments introduces in the mapper object.

check_arg_mapper(
  full_data,
  filter_values,
  type_covering,
  distance_type,
  clustering_type,
  linkage_type,
  optimal_clustering_mode = NA,
  silhouette_threshold = 0.25,
  dim_reduction,
  na.rm = TRUE
)

Arguments

full_data

Matrix with the columns of the input matrix corresponding to the individuals belonging to the level.

filter_values

Vector obtained after applying the filtering function to the input matrix, i.e, a vector with the filtering function values for each included sample.

type_covering

Parameter to choose how to construct the covering. Choose between "classical" or "uniform". "uniform" default option.

distance_type

Type of distance to be used for clustering. Choose between correlation ("correlation") and euclidean ("euclidean"). "correlation" default option.

clustering_type

Type of clustering method. Choose between "hierarchical" and "PAM" (“partition around medoids”) options. "hierarchical" default option.

linkage_type

Linkage criteria used in hierarchical clustering. Choose between "single" for single-linkage clustering, "complete" for complete-linkage clustering, "average" for average linkage clustering (or UPGMA) or "ward.D" for Ward's method. Only necessary for hierarchical clustering. "ward.D" default option.

optimal_clustering_mode

Method for selection optimal number of clusters. It is only necessary if the chosen type of algorithm is hierarchical. In this case, choose between "standard" (the method used in the original mapper article) or "silhouette". In the case of the PAM algorithm, the method will always be "silhouette".

silhouette_threshold

Minimum value of \(\overline{s}\) that a set of clusters must have to be chosen as optimal. Within each interval of the filter function, the average silhouette values \(\overline{s}\) are computed for all possible partitions from $2$ to $n-1$, where $n$ is the number of samples within a specific interval. The $n$ that produces the highest value of \(\overline{s}\) and that exceeds a specific threshold is selected as the optimum number of clusters. If no partition produces an \(\overline{s}\) exceeding the chosen threshold, all samples are then assigned to a unique cluster. The default value is $0.25$. The threshold of $0.25$ for \(\overline{s}\) has been chosen based on standard practice, recognizing it as a moderate value that reflects adequate separation and cohesion within clusters.

dim_reduction

Boolean. Parameter indicating whether or not centering, scaling and dimensionality reduction of the data is wanted prior to partial clustering in Mapper. This process is performed on each subset composed by the columns of the individuals that are part of each interval. TRUE indicates that centering, scaling and dimensionality reduction will be performed. FALSE means that this step will be omitted and clustering will be performed without transforming the data.

na.rm

logical. If TRUE, NA rows are omitted. If FALSE, an error occurs in case of NA rows.

Value

list with full_data, filter_values and optimal_clustering_mode